move external session migration helpers under config
