+++ codex-rs/app-server/src/config/external_agent_sessions/detect.rs
@@ -90,7 +90,7 @@ fn is_recent_enough(now: i64, latest_timestamp: i64) -> bool {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::external_agent_sessions::ledger::record_imported_session;
+    use crate::config::external_agent_sessions::ledger::record_imported_session;
     use codex_protocol::ThreadId;
     use serde_json::Value as JsonValue;
     use std::path::Path;
