+++ codex-rs/app-server/src/config/mod.rs
@@ -1,1 +1,2 @@
 pub(crate) mod external_agent_config;
+pub(crate) mod external_agent_sessions;
