+++ codex-rs/app-server/src/external_agent_config_api.rs
@@ -3,10 +3,10 @@ use crate::config::external_agent_config::ExternalAgentConfigMigrationItem as Co
 use crate::config::external_agent_config::ExternalAgentConfigMigrationItemType as CoreMigrationItemType;
 use crate::config::external_agent_config::ExternalAgentConfigService;
 use crate::config::external_agent_config::PendingPluginImport;
+use crate::config::external_agent_sessions::has_current_session_been_imported;
+use crate::config::external_agent_sessions::record_imported_session;
 use crate::error_code::internal_error;
 use crate::error_code::invalid_params;
-use crate::external_agent_sessions::has_current_session_been_imported;
-use crate::external_agent_sessions::record_imported_session;
 use codex_app_server_protocol::ExternalAgentConfigDetectParams;
 use codex_app_server_protocol::ExternalAgentConfigDetectResponse;
 use codex_app_server_protocol::ExternalAgentConfigImportParams;
