+++ codex-rs/app-server/src/lib.rs
@@ -79,7 +79,6 @@ mod device_key_api;
 mod dynamic_tools;
 mod error_code;
 mod external_agent_config_api;
-mod external_agent_sessions;
 mod filters;
 mod fs_api;
 mod fs_watch;
